requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
pymongo==4.8.0
python-dotenv==1.0.1
werkzeug==3.0.3
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
pymongo==4.7.2
Werkzeug==3.0.3
python-dotenv==1.0.1
